[update]schema.rb

+++ db/migrate/20210207083428_devise_create_users.rb
@@ -31,17 +31,17 @@ class DeviseCreateUsers < ActiveRecord::Migration[5.2]
       # t.integer  :failed_attempts, default: 0, null: false # Only if lock strategy is :failed_attempts
       # t.string   :unlock_token # Only if unlock strategy is :email or :both
       # t.datetime :locked_at
-      
-      
+
+
       t.string :name, null: false
-      t.integer :image_id 
+      t.bigint :image_id
       t.string :age
-      t.string :personality 
+      t.string :personality
       t.string :self_introduction
       t.string :message
-      t.boolean :is_instructor, default: false, null: false 
-      
-      
+      t.boolean :is_instructor, default: false, null: false
+
+
       t.timestamps null: false
     end
