+++ db/migrate/20210207085817_create_reservations.rb
@@ -1,8 +1,8 @@
 class CreateReservations < ActiveRecord::Migration[5.2]
   def change
     create_table :reservations do |t|
-      t.integer :to_reservation_id
-      t.integer :from_reservation_id
+      t.bigint :to_reservation_id
+      t.bigint :from_reservation_id
       t.datetime :date, null: false
       t.datetime :begin_time, null: false
       t.timestamps
