+++ db/migrate/20210207090040_create_user_reviews.rb
@@ -1,8 +1,8 @@
 class CreateUserReviews < ActiveRecord::Migration[5.2]
   def change
     create_table :user_reviews do |t|
-      t.integer :to_user_review_id
-      t.integer :from_user_review_id
+      t.bigint :to_user_review_id
+      t.bigint :from_user_review_id
       t.timestamps
     end
   end
