+++ db/migrate/20210207090254_create_user_rooms.rb
@@ -1,8 +1,8 @@
 class CreateUserRooms < ActiveRecord::Migration[5.2]
   def change
     create_table :user_rooms do |t|
-      t.integer "user_id", null: false
-      t.integer "room_id", null: false
+      t.bigint "user_id", null: false
+      t.bigint "room_id", null: false
       t.timestamps
     end
   end
